+++ starter/slack-spring-boot/src/main/kotlin/io/hndrs/slack/broker/receiver/SL4JLoggingHandler.kt
@@ -4,13 +4,13 @@ import io.hndrs.slack.broker.command.CommandHandler
 import io.hndrs.slack.broker.command.SlashCommand
 import io.hndrs.slack.broker.event.EventHandler
 import io.hndrs.slack.broker.event.SlackEvent
-import io.hndrs.slack.broker.installation.InstallationReceiver
+import io.hndrs.slack.broker.installation.InstallationHandler
 import io.hndrs.slack.broker.store.team.Team
 import org.slf4j.LoggerFactory
 import org.springframework.core.Ordered
 import org.springframework.http.HttpHeaders
 
-class SL4JLoggingHandler : EventHandler, CommandHandler, InstallationReceiver {
+class SL4JLoggingHandler : EventHandler, CommandHandler, InstallationHandler {
     /**
      * receivers will be sorted ascending by this order
      */